# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "abacus",
    "description": "",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/elections/{election_id}/apportionment": {
      "post": {
        "summary": "Get the apportionment for an election (coordinator_gsb)",
        "operationId": "election_apportionment",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Election Apportionment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionApportionmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Election apportionment is not yet available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Election apportionment is not possible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
...
    "/api/log": {
      "get": {
        "summary": "List audit events (administrator, coordinator_gsb)",
        "operationId": "audit_log_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, default 1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by log level",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "Filter by event type",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "user",
            "in": "query",
            "description": "Filter by user ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only show events since the specified timestamp",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log event list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "administrator",
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
    "/api/log-users": {
      "get": {
        "summary": "Lists all users that appear in the audit log (administrator, coordinator_gsb)",
        "operationId": "audit_log_list_users",
        "responses": {
          "200": {
            "description": "Audit log list of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogUser"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "administrator",
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogEvent": {
        "type": "object",
        "required": [
          "id",
          "time",
          "event",
          "event_name",
          "event_level",
          "ip"
        ],
        "properties": {
          "event": {},
          "event_level": {
            "$ref": "#/components/schemas/AuditEventLevel"
          },
          "event_name": {
            "$ref": "#/components/schemas/AuditEventType"
          },
          "id": {
            "$ref": "#/components/schemas/AuditLogEventId"
          },
          "ip": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "user_fullname": {
            "type": "string"
          },
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          },
          "user_role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AuditLogListResponse": {
        "type": "object",
        "required": [
          "events",
          "page",
          "pages",
          "per_page"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogEvent"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "AuditLogUser": {
        "type": "object",
        "required": [
          "id",
          "username",
          "fullname",
          "role"
        ],
        "properties": {
          "fullname": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CandidateNomination": {
        "type": "object",
        "required": [
          "preference_threshold",
          "chosen_candidates",
          "list_candidate_nomination"
        ],
        "properties": {
          "chosen_candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChosenCandidate"
            }
          },
          "list_candidate_nomination": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListCandidateNomination"
            }
          },
          "preference_threshold": {
            "$ref": "#/components/schemas/PreferenceThreshold"
          }
        }
      },
...
      "ChosenCandidate": {
        "type": "object",
        "description": "Chosen candidate",
        "required": [
          "number",
          "initials",
          "last_name",
          "locality",
          "list_number",
          "list_name"
        ],
        "properties": {
          "country_code": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/CandidateGender"
          },
          "initials": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "last_name_prefix": {
            "type": "string"
          },
          "list_name": {
            "type": "string"
          },
          "list_number": {
            "$ref": "#/components/schemas/PGNumber"
          },
          "locality": {
            "type": "string"
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "ElectionApportionmentResponse": {
        "type": "object",
        "required": [
          "seat_assignment",
          "candidate_nomination",
          "election_summary"
        ],
        "properties": {
          "candidate_nomination": {
            "$ref": "#/components/schemas/CandidateNomination"
          },
          "election_summary": {
            "$ref": "#/components/schemas/ElectionSummary"
          },
          "seat_assignment": {
            "$ref": "#/components/schemas/SeatAssignment"
          }
        }
      },
...
      "ElectionSummary": {
        "type": "object",
        "description": "Contains a summary of the election results, added up from the votes of all polling stations.",
        "required": [
          "voters_counts",
          "votes_counts",
          "differences_counts",
          "political_group_votes",
          "polling_station_investigations"
        ],
        "properties": {
          "differences_counts": {
            "$ref": "#/components/schemas/SummaryDifferencesCounts",
            "description": "The differences between voters and votes"
          },
          "political_group_votes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoliticalGroupCandidateVotes"
            },
            "description": "The summary votes for each political group (and each candidate within)"
          },
          "polling_station_investigations": {
            "$ref": "#/components/schemas/PollingStationInvestigations",
            "description": "Polling stations where results were investigated by the GSB"
          },
          "voters_counts": {
            "$ref": "#/components/schemas/VotersCounts",
            "description": "The total number of voters"
          },
          "votes_counts": {
            "$ref": "#/components/schemas/VotesCounts",
            "description": "The total number of votes"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorReference": {
        "type": "string",
        "description": "Error reference used to show the corresponding error message to the end-user",
        "enum": [
          "AirgapViolation",
          "AlreadyInitialised",
          "ApportionmentAllListsExhausted",
          "ApportionmentCommitteeSessionNotCompleted",
          "ApportionmentDrawingOfLotsRequired",
          "ApportionmentZeroVotesCast",
          "CommitteeSessionPaused",
          "DatabaseError",
          "DataEntryAlreadyClaimed",
          "DataEntryAlreadyFinalised",
          "DataEntryCannotBeDeleted",
          "DataEntryGetNotAllowed",
          "DataEntryNotAllowed",
          "EmlImportError",
          "EmlError",
          "EntryNotFound",
          "EntryNotUnique",
          "Forbidden",
          "InternalServerError",
          "InvalidCommitteeSessionStatus",
          "InvalidData",
          "InvalidHash",
          "InvalidJson",
          "InvalidPassword",
          "InvalidPoliticalGroup",
          "InvalidSession",
          "InvalidStateTransition",
          "InvalidUsernameOrPassword",
          "InvalidVoteCandidate",
          "InvalidVoteGroup",
          "InvalidXml",
          "InvestigationHasDataEntryOrResult",
          "InvestigationRequiresCorrectedResults",
          "NotInitialised",
          "OwnAccountCannotBeDeleted",
          "PasswordRejectionSameAsOld",
          "PasswordRejectionSameAsUsername",
          "PasswordRejectionTooShort",
          "PdfGenerationError",
          "PollingStationRepeated",
          "PollingStationValidationErrors",
          "RequestPayloadTooLarge",
          "Unauthorized",
          "UsernameNotUnique",
          "UserNotFound"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Response structure for errors",
        "required": [
          "error",
          "fatal",
          "reference"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "fatal": {
            "type": "boolean"
          },
          "reference": {
            "$ref": "#/components/schemas/ErrorReference"
          }
        },
        "additionalProperties": false
      },
...
      "ListCandidateNomination": {
        "type": "object",
        "required": [
          "list_number",
          "list_name",
          "list_seats",
          "preferential_candidate_nomination",
          "other_candidate_nomination",
          "updated_candidate_ranking"
        ],
        "properties": {
          "list_name": {
            "type": "string"
          },
          "list_number": {
            "$ref": "#/components/schemas/PGNumber"
          },
          "list_seats": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "other_candidate_nomination": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateVotes"
            }
          },
          "preferential_candidate_nomination": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateVotes"
            }
          },
          "updated_candidate_ranking": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Candidate"
            }
          }
        }
      },
...
      "PoliticalGroupCandidateVotes": {
        "type": "object",
        "required": [
          "number",
          "total",
          "candidate_votes"
        ],
        "properties": {
          "candidate_votes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateVotes"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "PreferenceThreshold": {
        "type": "object",
        "required": [
          "percentage",
          "number_of_votes"
        ],
        "properties": {
          "number_of_votes": {
            "$ref": "#/components/schemas/DisplayFraction"
          },
          "percentage": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "administrator",
          "coordinator_gsb",
          "coordinator_csb",
          "typist_gsb",
          "typist_csb"
        ]
      },
...
      "SeatAssignment": {
        "type": "object",
        "required": [
          "seats",
          "full_seats",
          "residual_seats",
          "quota",
          "steps",
          "final_standing"
        ],
        "properties": {
          "final_standing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListSeatAssignment"
            }
          },
          "full_seats": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "quota": {
            "$ref": "#/components/schemas/DisplayFraction"
          },
          "residual_seats": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "seats": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatChangeStep"
            }
          }
        }
      },
...
      "SummaryDifferencesCounts": {
        "type": "object",
        "description": "Contains a summary of the differences, containing which polling stations had differences.",
        "required": [
          "more_ballots_count",
          "fewer_ballots_count"
        ],
        "properties": {
          "fewer_ballots_count": {
            "$ref": "#/components/schemas/SumCount"
          },
          "more_ballots_count": {
            "$ref": "#/components/schemas/SumCount"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ABACUS_SESSION"
      }
    }
  }
}